+++ lib/responseDownloadFormats/html/css/compile.js
@@ -7,5 +7,5 @@ shell.exec(
 
 const css = "export const css = `" + fs.readFileSync("./css/output.css", "utf8") + "`;\n";
 
-fs.writeFileSync("./css/styles.ts", css);
+fs.writeFileSync("./css/compiled.ts", css);
 fs.unlinkSync("./css/output.css");
